Update pillar.example
Added version (explicit an older) so that the user is forced to think about this and sets the right value.
At least I was stuck finding out that in defaults.yaml it was set :) and reconfigured it from the example. Thanks for the formular trying to get it to run atm.